archimatch_app/serializers/ReceiptSerializer.py
from rest_framework import serializers
from archimatch_app.models import Receipt

class ReceiptSerializer(serializers.ModelSerializer):
    class Meta:
        model = Receipt
        fields = '__all__'