requirements.txt
asgiref==3.7.2
attrs==23.2.0
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==4.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
idna==3.6
importlib-resources==6.1.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.4
openapi-codec==1.3.2
phonenumbers==8.13.28
pillow==10.2.0
pkgutil_resolve_name==1.3.10
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.17.1
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
zipp==3.17.0
